BE-9: Migrate Beat Editor codebase to TypeScript
