+++ beat-editor/package.json
@@ -34,7 +34,12 @@
     ]
   },
   "devDependencies": {
+    "@types/lodash": "^4.17.20",
+    "@types/node": "^24.2.1",
+    "@types/react": "^19.1.9",
+    "@types/react-dom": "^19.1.7",
     "@vitejs/plugin-react": "^5.0.0",
+    "typescript": "^5.9.2",
     "vite": "^7.1.1",
     "vitest": "^3.2.4"
   }
