--- beat-editor/src/App.jsx
@@ -1,13 +1,0 @@
-import BeatChartContainer from "./components/BeatChart/BeatChartContainer";
-import '@fortawesome/fontawesome-free/css/all.min.css';
-import './styles.scss';
-
-function App() {
-  return (
-    <div className="container">
-      <BeatChartContainer />
-    </div>
-  );
-}
-
-export default App;
