--- beat-editor/src/constants/constants.js
@@ -1,7 +1,0 @@
-/**
- * EDIT TYPES
- */
-
-export const EDIT_TYPE_ADD = "ADD";
-export const EDIT_TYPE_DELETE = "DELETE";
-export const EDIT_TYPE_UNUSABLE = "UNUSABLE";
