+++ beat-editor/src/constants/constants.ts
@@ -1,0 +1,7 @@
+/**
+ * EDIT TYPES
+ */
+
+export const EDIT_TYPE_ADD = "ADD";
+export const EDIT_TYPE_DELETE = "DELETE";
+export const EDIT_TYPE_UNUSABLE = "UNUSABLE";
