--- beat-editor/src/index.jsx
@@ -1,9 +1,0 @@
-import React from "react";
-import ReactDOM from 'react-dom/client';
-import App from "./App";
-
-const e = document.getElementById('root');
-const root = ReactDOM.createRoot(e);
-
-root.render(<App />);
-
