+++ beat-editor/src/index.tsx
@@ -1,0 +1,8 @@
+import ReactDOM from 'react-dom/client';
+import App from "./App";
+
+const e = document.getElementById('root');
+const root = ReactDOM.createRoot(e as HTMLElement);
+
+root.render(<App />);
+
